% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imgs.R
\name{get_imgs}
\alias{get_imgs}
\alias{get_img_links}
\title{Get images or image links from page}
\usage{
get_imgs(page, url)

get_img_links(page, url)
}
\description{
Get images or image links from page
}
\keyword{internal}
