% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_df.R
\name{html_df}
\alias{html_df}
\title{Get a tabular summary of webpage content from a vector of urls}
\usage{
html_df(urlx)
}
\arguments{
\item{urlx}{A character vector containing urls}
}
\value{
A tibble with columns \code{url} the original vector of urls;
\code{title} the page title' \code{url2} the final url if the original was a
redirect; \code{page} the result of a \code{GET} command; \code{html} the
html content.
}
\description{
From a vector of urls, \code{html_df} will attempt to fetch the html,
get the page title and the language, returning the result as a dataframe.
}
\examples{
urlx <- c("https://ropensci.org/blog/2020/02/21/ropensci-leadership/",
          "https://es.wikipedia.org/wiki/Wikipedia_en_espa\%C3\%B1ol")
html_df(urlx)

}
\author{
Alastair Rushworth
}
