% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_df.R
\name{html_df}
\alias{html_df}
\title{Get a tabular summary of webpage content from a vector of urls}
\usage{
html_df(urlx)
}
\arguments{
\item{urlx}{A character vector containing urls}
}
\value{
A tibble with columns 
\itemize{
\item \code{url} the original vector of urls
\item \code{title} the page title
\item \code{lang} the langauge of the webpage
\item \code{url2} the final url (for example, if the original was a redirect) 
\item \code{images} list column containing page images
\item \code{twitter} comma separated string containing embedded links to twitter profiles 
\item \code{github} comma separated string containing embedded links to github profiles 
\item \code{linkedin} comma separated string containing embedded links to linkedin profiles 
\item \code{page} list column containing results of a \code{GET} command; 
\item \code{html} list column containing page xml documents
}
}
\description{
From a vector of urls, \code{html_df} will attempt to fetch the html,
get the page title and the language, returning the result as a dataframe.
}
\examples{
urlx <- c("https://ropensci.org/blog/2020/02/21/ropensci-leadership/",
          "https://es.wikipedia.org/wiki/Wikipedia_en_espa\%C3\%B1ol")
html_df(urlx)

}
\author{
Alastair Rushworth
}
