% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_df.R
\name{html_df}
\alias{html_df}
\title{Get a tabular summary of webpage content from a vector of urls}
\usage{
html_df(urlx, max_size = 5e+06, time_out = 10, show_progress = TRUE)
}
\arguments{
\item{urlx}{A character vector containing urls.}

\item{max_size}{Maximum size in bytes of pages to attempt to parse, defaults to \code{5000000}.
This is to avoid reading very large base64 encoded pages can cause \code{read_html()} to hang.}

\item{time_out}{Time in seconds to wait for \code{httr::GET()} when fetching the page.  Defaults 
to 10 seconds.}

\item{show_progress}{Logical, defaults to \code{TRUE}. Whether to print progress 
messages to the console while download and parse is in progress.}
}
\value{
A tibble with columns 
\itemize{
\item \code{url} the original vector of urls provided
\item \code{title} the page title, if found
\item \code{lang} inferred page language
\item \code{url2} the fetched url this may be different to the original eg. if it was redirected
\item \code{rss} a list of embedded rss feeds found on the page
\item \code{images} list of tibbles containing image links found on the page
\item \code{social} list of tibbles containing twitter, linkedin and github user info found on page
\item \code{size} the size of the downloaded page in bytes
\item \code{server} the page server
\item \code{accessed} datetime when the page was accessed
\item \code{published} page publication or last updated date, if detected 
\item \code{generator} the page generator, if found
\item \code{source} list column containing page xml documents
}
}
\description{
From a vector of urls, \code{html_df} will attempt to fetch the html.  From the 
html, html_df will attempt to look for a page title, rss feeds, images, embedded social media
profile handles and other page metadata.  Page language is inferred using the package \code{cld3}
which uses Google's Compact Language Detector 3.
}
\examples{
urlx <- c("https://ropensci.org/blog/2020/02/21/ropensci-leadership/",
          "https://es.wikipedia.org/wiki/Wikipedia_en_espa\%C3\%B1ol")
html_df(urlx)

}
\author{
Alastair Rushworth
}
